[Fix] fin

+++ srcs/render/set_world.c
@@ -44,12 +44,12 @@ void	ft_wear_texture(t_world *world, t_raycast *rc, t_texture *tex, t_draw *dr)
 			tex->type = EA;
 	}
 	tex->wall_x -= floor(tex->wall_x);
-	dr->tex.x = (int)(tex->wall_x * world->tmlx->timg_texture[tex->type].w);
+	dr->tex.x = (int)(tex->wall_x * (double)world->tmlx->timg_texture[tex->type].w);
 	if ((rc->is_side == 0 && rc->ray.x > 0) \
 	|| (rc->is_side == 1 && rc->ray.y < 0))
 		dr->tex.x = world->tmlx->timg_texture[tex->type].w - dr->tex.x - 1;
 	tex->step = 1.0 * world->tmlx->timg_texture[tex->type].h  / dr->line_h;
-	tex->pos = (dr->start - world->screen_h / (double)(2 + dr->line_h) / 2) * tex->step;
+	tex->pos = (dr->start - world->screen_h / 2.0 + dr->line_h / 2.0) * tex->step;
 }
 
 void	ft_fill_buf(t_world *world, t_raycast *rc, t_texture *tex, size_t x)
